+++ docs/changes.rst
@@ -23,6 +23,7 @@ Not yet released.
 * :ref:`addon-weblate.discovery.discovery` now preserves :ref:`componentlists`.
 * Nicer formatting of :ref:`glossary` tooltips.
 * :http:get:`/api/components/(string:project)/(string:component)/` now includes information about linked component.
+* Improved :ref:`workflow-customization` configuration forms.
 
 **Bug fixes**